+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 - The Paper Doll and Change Skin button can now be freely repositioned by holding `Shift + Right-Click`. As a result, the `Paper Doll Position` option has been removed.
 - Improved Paper Doll rendering.
+- Removed the `Scan Resource Packs` option.
+- Updated mod credits.
 
 **Full Changelog**: https://github.com/BrandonItaly/Bedrock-Skins/commits/master